+++ src/components/DataTable/DefaultQuery/types.ts
@@ -1,4 +1,0 @@
-export type DefaultQueryProps = {
-    state: any;
-    dispatch: any;
-}