+++ src/components/DataTable/ODataQuery/types.ts
@@ -1,4 +1,0 @@
-export type ODataQueryProps = {
-    state: any;
-    dispatch: any;
-}