+++ src/components/DataTable/Pagination.tsx
@@ -1,5 +1,9 @@
 import React from 'react';
-import { PaginationProps } from './Pagination/types';
+
+export type PaginationProps = {
+    getCount?: Function;
+    defaultPageSize?: number;
+}
 
 /**
  * Pagination plugin. Must be used together with `paginationReducer`
@@ -8,5 +12,5 @@ import { PaginationProps } from './Pagination/types';
  * @returns `JSX.Element`
  */
 export const Pagination: React.FC<PaginationProps> = (props) => {
-    return (<React.Fragment {...props}></React.Fragment>)
+    return (<React.Fragment {...props}>Pagination</React.Fragment>)
 }